hotel/serializers.py
from rest_framework import serializers
from .models import Hotel, Facility


class FacilitySerializer(serializers.ModelSerializer):
    class Meta:
        model = Facility
        fields = '__all__'


class HotelSerializer(serializers.ModelSerializer):
    facility = FacilitySerializer(read_only=True)

    class Meta:
        model = Hotel
        fields = '__all__'
