requirements.txt
autopep8==1.5.3
Django==2.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.1.2
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
toml==0.10.1
